efw-protocols: add implementation for protocol about hardware information
This commit adds implementation for protocol about hardware information.
The batch of code is moved from internal transaction module to new
module.
