+++ libs/efw/runtime/src/input_ctl.rs
@@ -5,7 +5,8 @@ use glib::{Error, FileError};
 use core::card_cntr;
 use core::elem_value_accessor::ElemValueAccessor;
 
-use efw_protocols::transactions::{HwCap, EfwPhysInput, NominalLevel, HwInfo};
+use efw_protocols::transactions::{EfwPhysInput, NominalLevel};
+use efw_protocols::hw_info::*;
 
 pub struct InputCtl {
     phys_inputs: usize,
