+++ libs/efw/runtime/src/mixer_ctl.rs
@@ -7,7 +7,8 @@ use alsa_ctl_tlv_codec::items::DbInterval;
 use core::card_cntr;
 use core::elem_value_accessor::ElemValueAccessor;
 
-use efw_protocols::transactions::{HwInfo, EfwPlayback, EfwMonitor, HwCap, EfwHwCtl, HwCtlFlag};
+use efw_protocols::transactions::{EfwPlayback, EfwMonitor, EfwHwCtl, HwCtlFlag};
+use efw_protocols::hw_info::*;
 
 pub struct MixerCtl {
     playbacks: usize,
